feat: update challenge page structure
