--- src/views/Challenges/IndexView.vue
@@ -1,9 +1,0 @@
-<script lang="ts" setup>
-
-</script>
-
-<template>
-    <RouterView />
-</template>
-
-<style></style>
