+++ src/views/frontend-challenges/FrontendChallengesView.vue
@@ -1,0 +1,2 @@
+<script lang="ts" setup></script>
+<template><RouterView /></template>
